perf(verify): gate the canonical depth→slot assignment on batch width
THE MEASUREMENT

PR #552 made the batched-verify graph key canonical (depths descending paired
with slots ascending), collapsing the n=8 key space from 266 to 3. A/B on dgx2,
same binary, same session, ONE variable (the kill switch
ATLAS_NO_CANONICAL_VERIFY_KEY=1) — tok/s, higher is better:

   C  | canonical ON            | canonical OFF          | verdict
  ----+-------------------------+------------------------+------------------
    2 | 30.09                   | 30.83                  | costs -2.4%
    4 | 64.00 (round 7: 65.56)  | 68.11 (round 6, no it) | costs ~-3.7%
    8 | 110.63                  | 106.48                 | GAINS +3.9%
   16 | 203.50                  | 203.44                 | no effect
   32 | 291.50                  | 291.52                 | no effect
   64 | 387.62                  | 386.99                 | no effect
  128 | 477.55                  | 477.69                 | no effect

The canonical assignment pays for itself only where the key space actually
explodes. Measured key counts against VERIFY_BATCHED_GRAPH_CAP = 32: n=2 -> 2,
n=4 -> 10, n=8 -> 160-253, n=16 -> 1. At n=2 the key space was 2 and at n=4 it
was 10 — there is almost nothing to collapse, and forcing the assignment costs
more than it saves. Above width 8 the whole question is moot: D-Cut is off
there (dcut_width_cap), ks is uniform, and both arms reduce to "sort by slot"
— which is exactly why those four rungs move by <= 0.2% either way.

Hypothesis for the cost below 8, recorded but NOT load-bearing (the threshold
is the measurement, not the mechanism): the forced assignment makes the
two-launch batched GDN conv+WY fast path decline, i.e. n*(2k-1) launches per
layer instead of 2 — 768 vs 96 per step at n=2, k=4 over 48 GDN layers. PR
#553's rate telemetry under ATLAS_MTP_ACCEPT_DEBUG reports that rate directly.

THE GATE

verify_key::CANONICAL_KEY_MIN_WIDTH = 8 — ONE named threshold, documented with
the table above (the ssm_reserve.rs / adaptive_rung.rs house style of recording
the numbers that justify a constant). At batch widths >= it the canonical
assignment applies; below it verify_batch_order / verify_batch_permutation take
their canonical = false arm.

  * ATLAS_CANONICAL_KEY_MIN_WIDTH=<n> sweeps the threshold without a rebuild
    (VALUE-parsed once per process behind a OnceLock, like dcut_width_cap; 0 =
    canonical everywhere = plain #552, a value above the widest batch = never;
    unset or unparseable falls back to the constant).
  * ATLAS_NO_CANONICAL_VERIFY_KEY (PRESENCE) still disables it ENTIRELY and
    DOMINATES both the width and the override.

WHY THE GATE BELONGS AT THE ASSIGNMENT SEAM

#552's structure is the reason this is a small change: plan decides order AND
assignment, mtp_step takes order only, verify_e2 builds the key from whatever
order it is handed. So there is exactly ONE decision to gate, and the gate is
one function — verify_key::canonical_assignment(n) — which both seams ask and
neither re-derives. Sprinkling a width test at each call site would let the
order and the assignment disagree, which is precisely the failure the seam
split was built to make unrepresentable (verify_k4_batch_step pins
drafts + 1 == ks[i]).

The one subtlety the gate introduces is chunking: mtp_step re-applies the ORDER
per verify chunk, so it must ask the gate with the width plan gated on
(batchable_idxs.len(), captured as batch_n), never the chunk's. Today those
coincide in every ragged case — D-Cut only prunes at n <= 8 and an n <= 8 batch
of rows <= 4 is a single chunk under the 96-row budget — but relying on that
coincidence is exactly the kind of thing that rots, so the width is passed
explicitly.

BYTE-IDENTITY BELOW THE THRESHOLD

The whole point of the gate, so it is pinned rather than argued:
below_the_threshold_is_byte_identical_to_pre_canonical enumerates every ragged
shape D-Cut can emit at n = 1, 2, 4, 7 (rows 2..=4 per sequence, over slot sets
that are neither sorted nor contiguous — 16660 cases) and asserts both the
ASSIGNMENT and the KEY BYTES equal a pre-#552 reproduction built from first
principles: a stable sort_by_key(|(a, k)| (Reverse(k), slot)) with each member
keeping its own depth — literally what both call sites ran before #552, not a
re-read of the canonical = false arm.

D-Cut itself is untouched below the threshold: this reverts the ASSIGNMENT, not
the pruning. below_the_gate_the_pairing_is_legacy_but_the_pruning_is_kept
composes select -> gate -> verify_batch_order the way plan does and pins that
the n=2 ragged {4,3} multiset (R=7 against the uniform 8) survives while each
sequence keeps the depth its own confidence earned.

The D-Cut planner and the SSM row-order invariants are otherwise unchanged:
canonical_order_holds_both_dispatch_invariants,
each_depth_run_owns_a_consecutive_slot_block,
canonical_assignment_preserves_the_selected_row_total and
chunk_ranges_* all still pass unmodified.

TESTS (+9, 35 total across the two seams; all pure/CPU)

  * below_the_threshold_is_byte_identical_to_pre_canonical — the guarantee.
  * n4_gated_key_is_the_legacy_bytes_not_the_canonical_ones — a hard byte pin
    so the two arms cannot silently converge.
  * at_the_threshold_the_gate_selects_the_canonical_arm — #552's collapse still
    happens at n=8, where it was measured to pay.
  * the_gate_boundary_is_the_constant / min_width_is_the_documented_default.
  * min_width_override_parses + min_width_override_moves_the_boundary — the
    override moves the boundary end-to-end in KEY BYTES (=4 makes n=4
    canonical, =16 makes n=8 legacy, =0 is plain #552).
  * kill_switch_dominates_width_and_override.
  * below_the_gate_the_pairing_is_legacy_but_the_pruning_is_kept (D-Cut seam).
  * #552's n8_key_space_collapses_266_to_3 keeps passing unmodified.

verify_key.rs's tests move to a sibling verify_key_tests.rs via #[path] (the
ssm_reserve.rs idiom) so the file stays under the 500-line cap: 420 -> 267.

EXPECTED EFFECT (to be measured on dgx2 — not claimed here)

C=2 30.09 -> ~30.8, C=4 64.00 -> ~68, C=8 keeps 110.63, C>=16 unchanged.
Floors that must not regress: C=1 23.66, C=2 31.00, C=4 68.11, C=8 110.63,
C=16 203.50, C=32 291.52, C=64 387.62, C=128 477.69.

Gates: cargo fmt --all -- --check, ATLAS_SKIP_BUILD=1 CUDARC_CUDA_VERSION=13000
cargo clippy --workspace --tests, scripts/check-license-headers.sh (2238 valid,
0 invalid), cargo test -p spark-model --lib (623 passed) and cargo test
-p spark-server --bins (2291 passed) — all green.
