+++ crates/spark-model/src/speculative/verify_key.rs
@@ -63,12 +63,110 @@
 //! `=0` is NOT off) restores the pre-canonical behaviour: each sequence keeps
 //! its own confidence-chosen depth and the batch is sorted deepest-first,
 //! ssm-slot second.
+//!
+//! # The width gate: it only pays where the key space explodes
+//!
+//! Collapsing the key space is not free — forcing the assignment overrides
+//! D-Cut's confidence pairing and re-shapes the depth runs — and the key
+//! space only explodes at the TOP of D-Cut's width range. Measured key
+//! counts against `VERIFY_BATCHED_GRAPH_CAP` = 32: n=2 → 2, n=4 → 10,
+//! **n=8 → 160-253**, n=16 → 1. At n=2 and n=4 there is essentially nothing
+//! to collapse, and the A/B says so — see [`CANONICAL_KEY_MIN_WIDTH`], which
+//! is the ONE threshold and carries the table. Below it the pre-canonical
+//! assignment is restored BYTE-IDENTICALLY; at/above it the canonical
+//! assignment applies. [`canonical_assignment`] is the single gate; call
+//! sites never re-derive it.
 
 /// Canonical assignment ON unless `ATLAS_NO_CANONICAL_VERIFY_KEY` is present.
 /// Read once per process.
 pub fn canonical_verify_key_enabled() -> bool {
     static ON: std::sync::OnceLock<bool> = std::sync::OnceLock::new();
     *ON.get_or_init(|| std::env::var_os("ATLAS_NO_CANONICAL_VERIFY_KEY").is_none())
+}
+
+/// Batch WIDTH (sequences) at or above which the canonical depth→slot
+/// assignment is applied. Below it [`verify_batch_order`] /
+/// [`verify_batch_permutation`] take their `canonical = false` arm, which is
+/// the pre-canonical (pre-PR-#552) behaviour byte for byte: each sequence
+/// keeps its own confidence-chosen depth and the batch sorts deepest-first,
+/// ssm-slot second, ties on input index (a stable
+/// `sort_by_key(|(a, k)| (Reverse(k), slot))`, exactly what both call sites
+/// used before).
+///
+/// Default **8**, from a same-binary same-session A/B on dgx2 (ladder-38
+/// round 7, tip `e0b845f11`) with ONE variable — the kill switch
+/// `ATLAS_NO_CANONICAL_VERIFY_KEY=1`. tok/s, higher is better:
+///
+/// ```text
+///  C  | canonical ON            | canonical OFF          | verdict
+/// ----+-------------------------+------------------------+------------------
+///   2 | 30.09                   | 30.83                  | costs -2.4%
+///   4 | 64.00 (round 7: 65.56)  | 68.11 (round 6, no it) | costs ~-3.7%
+///   8 | 110.63                  | 106.48                 | GAINS +3.9%
+///  16 | 203.50                  | 203.44                 | no effect
+///  32 | 291.50                  | 291.52                 | no effect
+///  64 | 387.62                  | 386.99                 | no effect
+/// 128 | 477.55                  | 477.69                 | no effect
+/// ```
+///
+/// The shape of that table follows the key counts (module docs): the
+/// collapse pays exactly where the arrangement space is large. Above width 8
+/// the gate is inert in either direction — D-Cut is off there
+/// (`dcut_width_cap`), `ks` is uniform, and both arms reduce to "sort by
+/// slot" (pinned by `uniform_depths_are_identical_under_both_arms`), which
+/// is why the C>=16 rungs move by <= 0.2% either way.
+///
+/// Hypothesis for the cost below 8, recorded but NOT load-bearing for this
+/// threshold (the threshold is the measurement, not the mechanism): forcing
+/// the assignment makes the two-launch batched GDN conv+WY fast path decline
+/// more often, i.e. `n*(2k-1)` launches per layer instead of 2 — 768 vs 96
+/// per step at n=2, k=4 over 48 GDN layers. PR #553's rate telemetry under
+/// `ATLAS_MTP_ACCEPT_DEBUG` reports that decline rate directly.
+pub const CANONICAL_KEY_MIN_WIDTH: usize = 8;
+
+/// Sweep the threshold without a rebuild: `ATLAS_CANONICAL_KEY_MIN_WIDTH=<n>`
+/// (VALUE-parsed; 0 = canonical at every width, a value above the widest
+/// batch = never). Unset or unparseable ⇒ [`CANONICAL_KEY_MIN_WIDTH`].
+/// Parsed once per process, like `dcut_width_cap`.
+pub fn canonical_key_min_width() -> usize {
+    static N: std::sync::OnceLock<usize> = std::sync::OnceLock::new();
+    *N.get_or_init(|| min_width_from_env(std::env::var_os(ENV_MIN_WIDTH)))
+}
+
+/// The env var name, named once so the parser and its tests cannot drift.
+const ENV_MIN_WIDTH: &str = "ATLAS_CANONICAL_KEY_MIN_WIDTH";
+
+/// Pure parse of [`ENV_MIN_WIDTH`] — the I/O lives in
+/// [`canonical_key_min_width`] so the policy is testable without touching
+/// process env (which a `OnceLock` would latch anyway).
+fn min_width_from_env(raw: Option<std::ffi::OsString>) -> usize {
+    raw.and_then(|v| v.into_string().ok())
+        .and_then(|v| v.trim().parse::<usize>().ok())
+        .unwrap_or(CANONICAL_KEY_MIN_WIDTH)
+}
+
+/// **THE GATE** — the one decision "does this batch get the canonical
+/// depth→slot assignment?". Both seams ask this and nothing else:
+/// `mtp_dcut::plan` (which decides order AND assignment) and `mtp_step`
+/// (permutation only), each passing the FULL batch width so the two can
+/// never disagree — `plan` gates on `batchable.len()`, and a chunked
+/// dispatch must use that same width, not the chunk's.
+///
+/// `n` is the batch width in SEQUENCES. All logic lives in
+/// [`canonical_assignment_at`]; this is only the env binding (SBIO).
+pub fn canonical_assignment(n: usize) -> bool {
+    canonical_assignment_at(n, canonical_key_min_width(), canonical_verify_key_enabled())
+}
+
+/// The gate policy, with its two environment inputs INJECTED (SBIO): the
+/// resolved threshold and whether the kill switch is CLEAR. The kill switch
+/// dominates — once `ATLAS_NO_CANONICAL_VERIFY_KEY` is set, no width and no
+/// `ATLAS_CANONICAL_KEY_MIN_WIDTH` value can turn the assignment back on.
+///
+/// Split out because `OnceLock`-latched env cannot be moved from a test, and
+/// a policy nobody can exercise is a policy nobody has checked.
+fn canonical_assignment_at(n: usize, min_width: usize, kill_switch_clear: bool) -> bool {
+    kill_switch_clear && n >= min_width
 }
 
 /// Dispatch ORDER for one verify batch — the permutation only.
@@ -165,256 +263,5 @@ pub fn verify_graph_key(pairs: &[(u32, u32)], wy_tables_null: bool) -> Vec<u32> 
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// Build the key the dispatch would produce for a batch given as
-    /// `(slot, k)` pairs in the scheduler's arbitrary pre-sort order.
-    fn key_for(batch: &[(usize, usize)], canonical: bool) -> Vec<u32> {
-        let slots: Vec<usize> = batch.iter().map(|&(s, _)| s).collect();
-        let ks: Vec<usize> = batch.iter().map(|&(_, k)| k).collect();
-        let (order, depths) = verify_batch_order(&slots, &ks, canonical);
-        let pairs: Vec<(u32, u32)> = order
-            .iter()
-            .zip(&depths)
-            .map(|(&i, &k)| (slots[i] as u32, k as u32))
-            .collect();
-        verify_graph_key(&pairs, false)
-    }
-
-    /// THE DEFECT, pinned. The same depth multiset spread over the same slots
-    /// in different arrangements — exactly what D-Cut re-ranking produces
-    /// step to step — must collapse onto ONE key.
-    #[test]
-    fn same_multiset_different_arrangement_is_one_key() {
-        // Multiset {4,4,3,3} over slots {0,1,2,3}: 4!/(2!·2!) = 6 arrangements.
-        let arrangements: [[(usize, usize); 4]; 6] = [
-            [(0, 4), (1, 4), (2, 3), (3, 3)],
-            [(0, 4), (1, 3), (2, 4), (3, 3)],
-            [(0, 4), (1, 3), (2, 3), (3, 4)],
-            [(0, 3), (1, 4), (2, 4), (3, 3)],
-            [(0, 3), (1, 4), (2, 3), (3, 4)],
-            [(0, 3), (1, 3), (2, 4), (3, 4)],
-        ];
-        let keys: std::collections::HashSet<Vec<u32>> =
-            arrangements.iter().map(|a| key_for(a, true)).collect();
-        assert_eq!(keys.len(), 1, "6 arrangements must collapse to 1 key");
-        // And the ONE key is the canonical pairing: slots ascending, depths
-        // descending.
-        assert_eq!(
-            keys.into_iter().next().unwrap(),
-            vec![0, 4, 1, 4, 2, 3, 3, 3, u32::MAX]
-        );
-    }
-
-    /// The n=8 arithmetic from the module docs, executed: the three depth
-    /// multisets the profile observed produce 266 keys pre-canonicalization
-    /// and 3 after. Generated exhaustively so the count is computed, not
-    /// asserted from a comment.
-    #[test]
-    fn n8_key_space_collapses_266_to_3() {
-        // (count of depth 4, count of 3, count of 2) for each observed shape.
-        let multisets = [(5usize, 2usize, 1usize), (4, 4, 0), (6, 0, 2)];
-        let mut legacy: std::collections::HashSet<Vec<u32>> = Default::default();
-        let mut canon: std::collections::HashSet<Vec<u32>> = Default::default();
-        for &(c4, c3, c2) in &multisets {
-            let mut depths: Vec<usize> = Vec::new();
-            depths.extend(std::iter::repeat_n(4usize, c4));
-            depths.extend(std::iter::repeat_n(3usize, c3));
-            depths.extend(std::iter::repeat_n(2usize, c2));
-            assert_eq!(depths.len(), 8);
-            // Every distinct arrangement of this multiset over slots 0..8.
-            let mut perm: Vec<usize> = (0..8).collect();
-            permute(&mut perm, 0, &mut |p| {
-                let batch: Vec<(usize, usize)> = (0..8).map(|s| (s, depths[p[s]])).collect();
-                legacy.insert(key_for(&batch, false));
-                canon.insert(key_for(&batch, true));
-            });
-        }
-        assert_eq!(legacy.len(), 266, "pre-canonical arrangement count");
-        assert_eq!(canon.len(), 3, "one key per depth multiset");
-    }
-
-    /// Every permutation of `v`, in place.
-    fn permute(v: &mut Vec<usize>, i: usize, f: &mut impl FnMut(&[usize])) {
-        if i == v.len() {
-            f(v);
-            return;
-        }
-        for j in i..v.len() {
-            v.swap(i, j);
-            permute(v, i + 1, f);
-            v.swap(i, j);
-        }
-    }
-
-    /// Different multisets must NOT collide — a graph baked for one depth
-    /// shape replaying against another is the state-poisoning failure mode.
-    #[test]
-    fn different_multisets_have_different_keys() {
-        let a = key_for(&[(0, 4), (1, 4), (2, 3), (3, 3)], true);
-        let b = key_for(&[(0, 4), (1, 3), (2, 3), (3, 3)], true);
-        let c = key_for(&[(0, 4), (1, 4), (2, 4), (3, 3)], true);
-        assert_ne!(a, b);
-        assert_ne!(a, c);
-        assert_ne!(b, c);
-    }
-
-    /// A different SLOT SET at the same depth multiset must also differ: the
-    /// graph bakes those pool addresses.
-    #[test]
-    fn different_slot_sets_have_different_keys() {
-        let a = key_for(&[(0, 4), (1, 3)], true);
-        let b = key_for(&[(0, 4), (2, 3)], true);
-        assert_ne!(a, b);
-    }
-
-    /// The sentinel keeps a table-less capture from replaying a table-full
-    /// step.
-    #[test]
-    fn wy_table_presence_splits_the_key() {
-        let pairs = [(0u32, 4u32), (1, 3)];
-        assert_ne!(
-            verify_graph_key(&pairs, true),
-            verify_graph_key(&pairs, false)
-        );
-    }
-
-    /// The invariants the SSM batched arms depend on, asserted explicitly on
-    /// the canonical order: slots ASCENDING in batch order (the
-    /// consecutive-slot precondition of `ssm_batched_recurrent.rs` and
-    /// `trait_decode_batched_conv_gdn_multi.rs`) and depths NON-INCREASING
-    /// (equal depths form contiguous runs — `decode_batched_inner`'s run
-    /// loop).
-    #[test]
-    fn canonical_order_holds_both_dispatch_invariants() {
-        // Deliberately hostile input: slot order and depth order disagree.
-        let slots = [7usize, 2, 5, 0, 3];
-        let ks = [2usize, 4, 2, 3, 4];
-        let (order, depths) = verify_batch_order(&slots, &ks, true);
-        let placed: Vec<usize> = order.iter().map(|&i| slots[i]).collect();
-        assert!(
-            placed.windows(2).all(|w| w[0] < w[1]),
-            "slots must be ascending in batch order, got {placed:?}"
-        );
-        assert!(
-            depths.windows(2).all(|w| w[0] >= w[1]),
-            "depths must be non-increasing, got {depths:?}"
-        );
-        // The multiset — and therefore Σ rows, the row budget and chunking —
-        // is preserved exactly.
-        let mut before = ks.to_vec();
-        before.sort_unstable();
-        let mut after = depths.clone();
-        after.sort_unstable();
-        assert_eq!(before, after);
-        assert_eq!(placed, vec![0, 2, 3, 5, 7]);
-        assert_eq!(depths, vec![4, 4, 3, 2, 2]);
-    }
-
-    /// Contiguous slots stay contiguous per depth RUN — the precondition the
-    /// two-launch conv+WY fast path actually checks.
-    #[test]
-    fn each_depth_run_owns_a_consecutive_slot_block() {
-        let slots = [0usize, 1, 2, 3, 4, 5, 6, 7];
-        let ks = [2usize, 4, 3, 4, 2, 3, 4, 3];
-        let (order, depths) = verify_batch_order(&slots, &ks, true);
-        let placed: Vec<usize> = order.iter().map(|&i| slots[i]).collect();
-        let mut g0 = 0usize;
-        while g0 < depths.len() {
-            let mut g1 = g0 + 1;
-            while g1 < depths.len() && depths[g1] == depths[g0] {
-                g1 += 1;
-            }
-            assert!(
-                placed[g0..g1].windows(2).all(|w| w[1] == w[0] + 1),
-                "run {g0}..{g1} (k={}) must be consecutive slots, got {:?}",
-                depths[g0],
-                &placed[g0..g1]
-            );
-            g0 = g1;
-        }
-    }
-
-    /// The permutation-only entry point must NEVER re-pair depths — a
-    /// downstream stage re-assigning them could hand a sequence a depth
-    /// deeper than the drafts the planner truncated it to.
-    #[test]
-    fn permutation_leaves_depths_attached_to_their_member() {
-        // An arrangement the canonical planner would never emit (depths
-        // ascending along the slots), to prove the permutation does not
-        // "repair" it.
-        let slots = [7usize, 2, 5, 0];
-        let ks = [4usize, 2, 3, 2];
-        for canonical in [true, false] {
-            let order = verify_batch_permutation(&slots, &ks, canonical);
-            let (_, assigned) = verify_batch_order(&slots, &ks, canonical);
-            let carried: Vec<usize> = order.iter().map(|&i| ks[i]).collect();
-            if canonical {
-                // The planner's entry point DOES re-pair; the permutation
-                // does not. That difference is the point of the split.
-                assert_eq!(carried, vec![2, 2, 3, 4]);
-                assert_eq!(assigned, vec![4, 3, 2, 2]);
-            } else {
-                assert_eq!(carried, assigned);
-            }
-        }
-    }
-
-    /// Idempotence: a chunked caller re-applies the ordering per chunk.
-    #[test]
-    fn canonical_order_is_idempotent() {
-        let slots = [7usize, 2, 5, 0, 3];
-        let ks = [2usize, 4, 2, 3, 4];
-        let (o1, d1) = verify_batch_order(&slots, &ks, true);
-        let s1: Vec<usize> = o1.iter().map(|&i| slots[i]).collect();
-        let (o2, d2) = verify_batch_order(&s1, &d1, true);
-        assert_eq!(o2, (0..s1.len()).collect::<Vec<_>>());
-        assert_eq!(d2, d1);
-    }
-
-    /// Kill switch: the legacy ordering keeps each sequence's own depth, so
-    /// the same multiset in different arrangements yields DIFFERENT keys —
-    /// today's behaviour, restored exactly.
-    #[test]
-    fn kill_switch_restores_the_arrangement_keyed_behaviour() {
-        let a = key_for(&[(0, 4), (1, 4), (2, 3), (3, 3)], false);
-        let b = key_for(&[(0, 3), (1, 3), (2, 4), (3, 4)], false);
-        assert_ne!(a, b, "legacy keys must still separate arrangements");
-        // Legacy order is deepest-first, slot second — each member keeps its
-        // OWN depth.
-        assert_eq!(b, vec![2, 4, 3, 4, 0, 3, 1, 3, u32::MAX]);
-    }
-
-    /// Uniform depths (D-Cut off, above `dcut_width_cap`, or ratio 1.0):
-    /// both arms reduce to "sort by slot", so the whole D-Cut-off regime is
-    /// byte-identical under either setting.
-    #[test]
-    fn uniform_depths_are_identical_under_both_arms() {
-        let slots = [3usize, 1, 2, 0];
-        let ks = [3usize; 4];
-        let canon = verify_batch_order(&slots, &ks, true);
-        let legacy = verify_batch_order(&slots, &ks, false);
-        assert_eq!(canon, legacy);
-        assert_eq!(
-            canon.0.iter().map(|&i| slots[i]).collect::<Vec<_>>(),
-            vec![0, 1, 2, 3]
-        );
-    }
-
-    /// Env-independent as long as the test process does not set the kill
-    /// switch (CI does not) — same pattern as `dcut_width_cap`'s default pin.
-    #[test]
-    fn canonical_is_on_by_default() {
-        assert!(canonical_verify_key_enabled());
-    }
-
-    /// Degenerate widths must not panic: the key path runs for every batch
-    /// the scheduler forms.
-    #[test]
-    fn empty_and_single_batches_are_well_formed() {
-        assert_eq!(verify_batch_order(&[], &[], true), (vec![], vec![]));
-        assert_eq!(verify_batch_order(&[5], &[3], true), (vec![0], vec![3]));
-        assert_eq!(verify_graph_key(&[], false), vec![u32::MAX]);
-    }
-}
+#[path = "verify_key_tests.rs"]
+mod tests;
